+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Revision History
 ================
 
+0.0.2 (2018/09/01)
+------------------
+
+- Words embedding bias
+
+  - Generating analogies along the bias direction
+  - Standard evaluations of words embedding (word pairs and analogies)
+  - Plotting indirect bias
+  - Scatter plot of bias direction projections between two words embedding
+  - Improved verbose mode
+
+
 0.0.1 (2018/08/17)
 ------------------